additional changes to ensure case insensitivity
