+++ koku/api/provider/view.py
@@ -17,15 +17,11 @@
 
 """View for Providers."""
 import logging
-from functools import reduce
-from operator import and_
 
-from django.db.models import Q
 from django.shortcuts import get_object_or_404
 from django.utils.encoding import force_text
 from django.views.decorators.cache import never_cache
-from django_filters import CharFilter, FilterSet
-from django_filters.filters import BaseCSVFilter
+from django_filters import FilterSet
 from django_filters.rest_framework import DjangoFilterBackend
 from rest_framework import mixins, status, viewsets
 from rest_framework.exceptions import APIException
@@ -33,6 +29,7 @@ from rest_framework.permissions import AllowAny
 from rest_framework.response import Response
 from rest_framework.serializers import UUIDField
 
+from api.common.filters import CharListFilter
 from api.iam.models import Customer
 from api.provider import serializers
 from api.provider.models import Provider
@@ -41,18 +38,6 @@ from .provider_manager import ProviderManager, ProviderManagerError
 
 
 LOG = logging.getLogger(__name__)
-
-
-class CharListFilter(BaseCSVFilter, CharFilter):
-    """Add query filter capability to provide an anded list of filter values."""
-
-    def filter(self, qs, value):
-        """Filter to create a composite and filter of the value list."""
-        if not value:
-            return qs
-        value_list = ','.join(value).split(',')
-        queries = [Q(**{self.lookup_expr: val}) for val in value_list]
-        return qs.filter(reduce(and_, queries))
 
 
 class ProviderFilter(FilterSet):
@@ -142,14 +127,17 @@ class ProviderViewSet(mixins.CreateModelMixin,
     @never_cache
     def create(self, request, *args, **kwargs):
         """Create a Provider."""
-        request.data['type'] = request.data.get('type', '').lower()
+        provider_type = request.data.get('type', '')
+        if provider_type and Provider.PROVIDER_CASE_MAPPING.get(provider_type.lower()):
+            request.data['type'] = request.data.get('type', '').lower()
         return super().create(request=request, args=args, kwargs=kwargs)
 
     @never_cache
     def update(self, request, *args, **kwargs):
         """Update a Provider."""
-        if request.data.get('type'):
-            request.data['type'] = request.data.get('type').lower()
+        provider_type = request.data.get('type', '')
+        if provider_type and Provider.PROVIDER_CASE_MAPPING.get(provider_type.lower()):
+            request.data['type'] = provider_type.lower()
         if request.method == 'PATCH':
             raise ProviderMethodException('PATCH not supported')
         user = request.user
